Split `metric.py` into two modules
