+++ DiverseSelector/base.py
@@ -28,8 +28,13 @@ from pathlib import PurePath
 from typing import Union
 
 from DiverseSelector.feature import feature_reader
-from DiverseSelector.metric import compute_distance_matrix, entropy,\
-    gini_coefficient, logdet, shannon_entropy, total_diversity_volume, wdud
+from DiverseSelector.distance import compute_distance_matrix
+from DiverseSelector.diversity import (entropy,
+                                       gini_coefficient,
+                                       logdet,
+                                       shannon_entropy,
+                                       total_diversity_volume,
+                                       wdud)
 from DiverseSelector.utils import PandasDataFrame
 import numpy as np
 import pandas as pd
